renamed product types ctrl and refactored promise chain
